requirements.txt
Flask==2.3.3
Pillow==10.0.1
pytesseract==0.3.10
werkzeug==2.3.7
azure-cognitiveservices-vision-computervision==0.9.0
msrest==0.7.1
reportlab==4.0.4
gunicorn==21.2.0